requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0